+++ 01_主題課程/02_在 Discord 跟 AI 機器人聊天/practice/practice_3.py
@@ -5,44 +5,40 @@ import google.generativeai as genai
 
 from discord.ext import commands
 
-# 填入你的 Discord Token 與 gemini 的 API key
-DC_TOKEN = "YOUR DCBOT TOKEN"
-GEMINI_API_KEY = "YOUR GEMINI API KEY"
+# TODO : 填入你的 Discord Token 與 gemini 的 API key
+DC_TOKEN = "YOUR_DISCORD_TOKEN"
+GEMINI_API_KEY = "YOUR_GEMINI_API_KEY"
 
-# 啟動 gemini
 genai.configure(api_key=GEMINI_API_KEY)
 
-# 設定模型
-model = genai.GenerativeModel('gemini-2.0-flash')
+model = genai.GenerativeModel(
+    'gemini-2.0-flash',
+    # TODO : 請加入風格化輸出
+)
 
 
-# intents 為機器人的權限，這裡設定為全開
 intents = discord.Intents.all()
 
-# bot 是機器人的本體，這邊是設定他前綴，這邊設定為 %
 bot = commands.Bot(command_prefix='%', intents=intents)
 
+chat = # TODO : 初始化對話歷史
+
 @bot.event
-# 當機器人完成啟動
 async def on_ready():
     print(f"目前登入身份 --> {bot.user}")
 
-# TODO: 建立對應的 command
-# 要求：
-# 1. 利用 command 與 gemini 進行對話
-# 2. 要使用風格化輸出，請加入自己的創意
-# 3. 請加入預設的回答，防止 gemini 無法回答時，機器人回覆「我無法回答您的問題」
-# 4. 請實作記憶功能
-# [Start] 請在以下空間撰寫機器人程式
+@bot.command()
+async def gemini(ctx, prompt):
+    response = # TODO : 請加入相對應的對話函式
+    
+    text = response.text
 
+    if not text:
+        await ctx.send("我無法回答您的問題")
+        return
 
+    await ctx.send(text)
 
-
-
-
-# [End] 請在以上空間撰寫機器人程式
-
-# 啟動機器人
 async def main():
     await bot.start(DC_TOKEN)
